proxima_core_engine/requirements.txt
amqp==5.1.1
antlr4-python3-runtime==4.8
arabic-reshaper==3.0.0
asgiref==3.6.0
asn1crypto==1.5.1
async-timeout==4.0.2
attrs==22.2.0
autobahn==23.1.2
Automat==22.10.0
billiard==3.6.4.0
bitarray==2.7.3
boto3==1.26.88
botocore==1.29.88
celery==5.2.7
certifi==2022.12.7
cffi==1.15.1
channels==4.0.0
channels-redis==4.1.0
charset-normalizer==3.1.0
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
colorama==0.4.6
constantly==15.1.0
coreapi==2.3.3
coreschema==0.0.4
cron-descriptor==1.2.35
cryptography==40.0.1
cssselect2==0.7.0
Cython==0.29.33
daphne==4.0.0
Deprecated==1.2.13
dj-rest-auth==3.0.0
Django==4.1.7
django-cacheops==7.0
django-celery-beat==2.5.0
django-celery-results==2.5.0
django-cors-headers==3.14.0
django-extensions==3.2.1
django-oauth-toolkit==2.2.0
django-rest-swagger==2.2.0
django-ses==3.3.0
django-timezone-field==5.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
drf-yasg==1.21.5
exceptiongroup==1.1.0
filelock==3.10.0
funcy==1.18
gunicorn==20.1.0
html5lib==1.1
huggingface-hub==0.13.3
hydra-core==1.0.7
hyperlink==21.0.0
idna==3.4
incremental==22.10.0
inflection==0.5.1
iniconfig==2.0.0
install==1.3.5
itypes==1.2.0
Jinja2==3.1.2
jmespath==1.0.1
jwcrypto==1.4.2
kombu==5.2.4
lxml==4.9.2
MarkupSafe==2.1.2
mpmath==1.3.0
msgpack==1.0.5
networkx==3.0
oauthlib==3.2.2
omegaconf==2.0.6
openapi-codec==1.3.2
oscrypto==1.3.0
packaging==23.0
phonenumbers==8.13.7
Pillow==9.4.0
pluggy==1.0.0
portalocker==2.7.0
prompt-toolkit==3.0.38
psycopg2-binary==2.9.5
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pyHanko==0.17.2
pyhanko-certvalidator==0.20.1
PyJWT==2.6.0
pyOpenSSL==23.1.1
pypdf==3.8.0
pypng==0.20220715.0
pytest==7.2.0
pytest-asyncio==0.21.0
pytest-django==4.5.2
python-bidi==0.4.2
python-crontab==2.7.1
python-dateutil==2.8.2
python-dotenv==1.0.0
python-flutterwave==0.8.10
pytz==2022.7.1
pytz-deprecation-shim==0.1.0.post0
PyYAML==6.0
qrcode==7.4.2
redis==4.5.3
regex==2022.10.31
reportlab==3.6.12
requests==2.28.2
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.7
s3transfer==0.6.0
sacrebleu==2.3.1
service-identity==21.1.0
simplejson==3.18.4
six==1.16.0
soundfile==0.12.1
sqlparse==0.4.3
svglib==1.5.1
sympy==1.11.1
tablib==3.4.0
tabulate==0.9.0
tinycss2==1.2.1
tokenizers==0.13.2
tomli==2.0.1
torch==2.0.0
torchaudio==2.0.1
tqdm==4.65.0
transformers==4.27.2
Twisted==22.10.0
txaio==23.1.1
typing_extensions==4.5.0
tzdata==2022.7
tzlocal==4.3
uritemplate==4.1.1
uritools==4.0.1
urllib3==1.26.15
vine==5.0.0
wcwidth==0.2.6
webencodings==0.5.1
whitenoise==6.4.0
wrapt==1.15.0
xhtml2pdf==0.2.9
zope.interface==6.0
